move into one file
https://coveord.atlassian.net/browse/KIT-1725
