+++ packages/atomic/stencil.config.ts
@@ -62,7 +62,7 @@ export const config: Config = {
   namespace: 'atomic',
   taskQueue: 'async',
   sourceMap: true,
-  globalScript: 'src/global.ts',
+  globalScript: 'src/global/global.ts',
   outputTargets: [
     react({
       componentCorePackage: '@coveo/atomic',
